refactor(content): extract reusable content manager
-ContentTab.tsx was a 940-line private admin component. Split it into
components/content/ so a non-admin surface can reuse it, and add a scope
prop that hides the Created By column when a user is viewing only their
own rows.

Adds an inline status toggle and row selection with bulk publish, which
is what makes reviewing a large import tractable.
