+++ apps/www/src/routes/admin/_components/-UsersTab.tsx
@@ -56,7 +56,7 @@ import {
   useReplaceAdminUserSocialLinks,
   apiUrl
 } from '@/lib/http'
-import { ImageUploadField } from './-ImageUploadField'
+import { ImageUploadField } from '@/components/content/ImageUploadField'
 
 const ROLES = ['admin', 'editor', 'creator', 'user'] as const
 type UserRole = (typeof ROLES)[number]
